+++ components/Profile/AdminTeacherProfile/ProfilePicture.module.css
@@ -1,0 +1,39 @@
+.circleContainer {
+    display: flex;
+    justify-content: center;
+}
+
+.dropZone {
+    display: flex;
+    justify-content: center;
+}
+
+.dropZoneContentDisabled {
+    height: 240px;
+    width: 240px;
+    background-color: #bbb;
+    border-radius: 50%;
+    cursor: auto;
+}
+
+.dropZoneContent {
+    composes: dropZoneContentDisabled;
+    cursor: pointer;
+}
+
+.dropZoneImage {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+.dropZoneText {
+    color: red;
+}
+
+.circle {
+    height: 240px;
+    width: 240px;
+    background-color: #bbb;
+    border-radius: 50%;
+}
